+++ CHANGELOG.md
@@ -6,12 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2022-07-22
+## [1.1.0] - 2024-01-11
 ### Added
-- Initial release.
+- New `Util` static class with some helper functions.
+- A `examples/get_args.php` testing `Util::parse_get_args()` function.
+### Changed
+- `ParamsApp` now extends `Util` so will have the utility functions available.
+
+## [1.0.1] - 2022-07-22
+### Added
+- First release (replaced initially broken `1.0.0`).
 - Pulls all `Lum\CLI` libraries out of `lum-framework`.
 - Also pulls `Fakeserver` library out of `lum-core`.
 
-[Unreleased]: https://github.com/supernovus/lum.cli.php/compare/v1.0.0...HEAD
-[1.0.0]: https://github.com/supernovus/lum.cli.php/releases/tag/v1.0.0
+[Unreleased]: https://github.com/supernovus/lum.cli.php/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/supernovus/lum.cli.php/compare/v1.0.1...v1.1.0
+[1.0.1]: https://github.com/supernovus/lum.cli.php/releases/tag/v1.0.1